Refactor and update dashboard and routing structure
Updated the routing structure to use the new `MainDashboard` as the home screen and refactored the dashboard components to reside under `lib/screens/dashboard_screen`. Added new dashboard screens such as `AboutDashBoard`, `BlogListDashBoard`, `ServiceDashBoard`, `TeamDashBoard`, and `TestimonialsDashBoard`. Also, updated the `TabSidebar` and `Sidebar` to use the new dashboard screens and added a new `SiderMenu` for navigation.